refactor(compiler-shared): split types.ts into domain modules
Pure type move, no public surface change. The 1044-line types.ts becomes an
18-line barrel re-exporting types/{diagnostics,extraction,tokens,recipes,css,
spec,codegen,parsing,config,callbacks,build-info,compiler}.ts. All 125 exports
preserved; consumers import ./types unchanged.
