+++ packages/compiler-shared/src/types.ts
@@ -4,1041 +4,15 @@
  * both speak. Pure types, no runtime, no internal imports (leaf module).
  */
 
-export interface Span {
-  start: number
-  end: number
-}
-
-/** 1-indexed line, 1-indexed UTF-16 column — matches `tsc`/editor reporting. */
-export interface SourceLocation {
-  line: number
-  column: number
-}
-
-export interface SourceRange {
-  start: SourceLocation
-  end: SourceLocation
-}
-
-export interface DiagnosticLabel {
-  message?: string
-  /** UTF-8 byte offsets. */
-  span?: Span
-  location?: SourceRange
-}
-
-export type DiagnosticSeverity = 'info' | 'warning' | 'error'
-
-export interface Diagnostic {
-  code: string
-  message: string
-  severity: DiagnosticSeverity
-  file?: string
-  category?: string
-  /** UTF-8 byte offsets. */
-  span?: Span
-  location?: SourceRange
-  labels?: DiagnosticLabel[]
-  help?: string[]
-}
-
-export type MatchCategory = 'css' | 'recipe' | 'pattern' | 'jsx' | 'tokens'
-
-/** One atomic style declaration. Conditions are outer→inner; unconditional
- *  atoms have an empty array. */
-export interface Atom {
-  prop: string
-  value: string | number | boolean | null
-  conditions: string[]
-}
-
-export type UtilityValueInput = string | number | boolean | null
-
-export type UtilityResolvedScalar = string | number | boolean
-
-export type UtilityValueSource =
-  | { type: 'value-map'; key: string; aliases: string[] }
-  | { type: 'literal'; aliases: string[] }
-  | { type: 'token-reference' }
-  | { type: 'arbitrary' }
-
-export interface ResolveUtilityValueInput {
-  prop: string
-  value: UtilityValueInput
-}
-
-export interface ResolvedUtilityValue {
-  utility: string
-  className: string
-  cssValue: UtilityResolvedScalar
-  important: boolean
-  source: UtilityValueSource
-}
-
-/** A token that carries a given value — a candidate the developer can pick. */
-export interface TokenSuggestion {
-  /** Category-relative path (`red.500`, `fg.error`). */
-  token: string
-  /** `true` when the token is from the semantic layer. */
-  semantic: boolean
-  /** `true` when the token has condition variants (themes) — not a static equal. */
-  conditional: boolean
-}
-
-/** `recipe` is the serialized shape of `pandacss_recipes::Recipe`/`SlotRecipe`. */
-export interface RecipeEntry {
-  file: string
-  spanStart: number
-  recipe: unknown
-}
-
-export interface RecipeStyleEntry {
-  prop: string
-  value: string | number | boolean | null
-  conditions: string[]
-}
-
-export interface RecipeStyleGroup {
-  recipe: string
-  slot?: string | null
-  className: string
-  entries: RecipeStyleEntry[]
-}
-
-export interface EncodedRecipeStyles {
-  base: RecipeStyleGroup[]
-  variants: RecipeStyleGroup[]
-  atomic: Atom[]
-}
-
-export interface ParseFileReport {
-  /** Source path parsed for this report. */
-  path: string
-  /** Number of `css(...)` calls extracted from this file. */
-  cssCalls: number
-  /** Number of `cva(...)` recipe calls extracted from this file. */
-  cvaCalls: number
-  /** Number of `sva(...)` slot-recipe calls extracted from this file. */
-  svaCalls: number
-  /** Number of JSX style usages extracted from this file. */
-  jsxUsages: number
-  /** Diagnostics produced while parsing this file. */
-  diagnostics: Diagnostic[]
-}
-
-export interface ProjectSummary {
-  filesProcessed: number
-  atomCount: number
-  recipeCount: number
-  slotRecipeCount: number
-}
-
-export interface ParsedFileView {
-  path: string
-  atoms: Atom[]
-  diagnostics: Diagnostic[]
-  recipes: RecipeEntry[]
-  slotRecipes: RecipeEntry[]
-}
-
-export interface StaticPatternResult {
-  atoms: Atom[]
-  diagnostics: Diagnostic[]
-}
-
-export type CodegenArtifactId =
-  | 'conditions'
-  | 'css-index'
-  | 'cx'
-  | 'helpers'
-  | 'jsx-create-recipe-context'
-  | 'jsx-create-slot-recipe-context'
-  | 'jsx-factory'
-  | 'jsx-index'
-  | 'jsx-is-valid-prop'
-  | 'jsx-patterns'
-  | 'patterns'
-  | 'themes'
-  | 'types'
-
-export type CodegenDependency =
-  | 'outExtension'
-  | 'forceImportExtension'
-  | 'conditions'
-  | 'hash'
-  | 'jsxFactory'
-  | 'jsxFramework'
-  | 'jsxStyleProps'
-  | 'patterns'
-  | 'prefix'
-  | 'recipes'
-  | 'separator'
-  | 'syntax'
-  | 'themes'
-  | 'tokens'
-  | 'utilities'
-
-export interface GenerateArtifactOptions {
-  forceImportExtension?: boolean
-}
-
-export interface CodegenOptions extends GenerateArtifactOptions {
-  outdir?: string
-  cwd?: string
-}
-
-export interface WriteArtifactsOptions extends GenerateArtifactOptions {
-  outdir: string
-  cwd?: string
-  artifacts?: CodegenArtifact[]
-}
-
-export interface WriteCssOptions extends CompileOptions {
-  outfile: string
-  cwd?: string
-}
-
-/** CSS emit overrides shared by cssgen, build, and dev. */
-export interface CssOutputOptions {
-  /** Layer filter for split output. Omit to emit all layers. */
-  layers?: StylesheetLayerName[]
-  emitLayerDeclaration?: boolean
-  minify?: boolean
-}
-
-/** Options for {@link Compiler.getLayerCss} / {@link Driver.getLayerCss}. */
-export interface LayerCssOptions {
-  layers: StylesheetLayerName[]
-  emitLayerDeclaration?: boolean
-  minify?: boolean
-}
-
-/** Options for in-memory split CSS generation. */
-export type SplitCssOptions = CssOutputOptions
-
-export interface WriteLayerCssOptions extends LayerCssOptions {
-  outfile: string
-  cwd?: string
-}
-
-export interface WriteSplitCssOptions extends CssOutputOptions {
-  outdir?: string
-  cwd?: string
-}
-
-/** Scan overrides for `Compiler.scan`/`parseFiles`. Omitted fields fall back to the
- *  config's `include`/`exclude`/`cwd`. */
-export interface ScanOptions {
-  include?: string[]
-  exclude?: string[]
-  cwd?: string
-}
-
-/** Resolved cascade-layer names (config overrides merged over defaults). */
-export interface LayerNames {
-  reset: string
-  base: string
-  tokens: string
-  recipes: string
-  utilities: string
-}
-
-/** The five cascade layers, in emit order. */
-export type StylesheetLayerName = 'reset' | 'base' | 'tokens' | 'recipes' | 'utilities'
-
-/** One split CSS file, relative to `outdir`. */
-export interface CssFile {
-  path: string
-  code: string
-}
-
-/** A source glob with its static base directory (the dir a watcher subscribes to). */
-export interface SourceEntry {
-  base: string
-  pattern: string
-}
-
-export type UsageKind = 'token' | 'property' | 'recipe' | 'pattern' | 'keyframe'
-
-/** One classified Panda usage with its source range — for reporting, lint, IDE.
- *  `name` is a token path, canonical property, or recipe/pattern name. */
-export interface UsageSite {
-  kind: UsageKind
-  name: string
-  range: SourceRange
-}
-
-export interface SpecUtilityProperty {
-  name: string
-  cssProperty?: string
-  tokenCategory?: string
-  literals: string[]
-  alias: string
-}
-
-export interface SpecTokenCategory {
-  name: string
-  typeName: string
-  values: string[]
-}
-
-/** Tooling introspection snapshot — read once, index on the host (never query
- *  the engine per-item in a hot loop). Powers reporting / formatting / linting. */
-/** `true` = deprecated with no message; a string = the author's deprecation message. */
-export type Deprecation = true | string
-
-export interface SpecVariant {
-  values: string[]
-  allowsBoolean: boolean
-}
-
-export interface SpecRecipe {
-  name: string
-  typeName: string
-  variants: Record<string, SpecVariant>
-  deprecated?: Deprecation
-}
-
-export interface SpecSlotRecipe extends SpecRecipe {
-  slots: string[]
-}
-
-export interface SpecPattern {
-  name: string
-  typeName: string
-  strict: boolean
-  blocklist: string[]
-  properties: Record<string, unknown>
-  deprecated?: Deprecation
-}
-
-export interface Spec {
-  conditions: { keys: string[]; breakpoints: string[]; containers: string[] }
-  tokens: {
-    categories: Record<string, SpecTokenCategory>
-    colorPalettes: string[]
-    /** `path -> value` (empty value means it equals the token's CSS var). */
-    values: Record<string, string>
-    /** `token path -> deprecation` (`true` or an author message). */
-    deprecated: Record<string, Deprecation>
-  }
-  utilities: {
-    properties: Record<string, SpecUtilityProperty>
-    /** `shorthand -> canonical property`. */
-    shorthands: Record<string, string>
-    /** `property -> deprecation` (always `true`; utility config has no message). */
-    deprecated: Record<string, Deprecation>
-  }
-  /** Keyed by pattern name. */
-  patterns: Record<string, SpecPattern>
-  /** Keyed by recipe name. */
-  recipes: Record<string, SpecRecipe>
-  /** Keyed by slot-recipe name. */
-  slotRecipes: Record<string, SpecSlotRecipe>
-  /** Canonical emit order for property names (for a stable property sort). */
-  propertyOrder: string[]
-  jsxFactory?: string
-  /** Normalized import map. */
-  importMap?: ImportMapOutput
-}
-
-export interface CodegenFile {
-  path: string
-  code: string
-  dependencies: CodegenDependency[]
-}
-
-export interface CodegenArtifact {
-  id: CodegenArtifactId
-  files: CodegenFile[]
-}
-
-export interface CompileFileManifest {
-  path: string
-  hash: string
-}
-
-export interface CompileManifest {
-  files: CompileFileManifest[]
-  tokens: string[]
-}
-
-export interface CompileLayerRange {
-  start: number
-  end: number
-}
-
-export interface CompileLayerRanges {
-  reset?: CompileLayerRange
-  base?: CompileLayerRange
-  tokens?: CompileLayerRange
-  recipes?: CompileLayerRange
-  utilities?: CompileLayerRange
-}
-
-export interface CompileOutput {
-  css: string
-  sourceMap?: string
-  manifest: CompileManifest
-  layerRanges: CompileLayerRanges
-  diagnostics: Diagnostic[]
-}
-
-export interface WriteCssResult extends CompileOutput {
-  path: string
-}
-
-export interface WriteFilesResult {
-  root: string
-  paths: string[]
-  files: CssFile[]
-}
-
-export interface CompileOptions {
-  /** Emit Panda's leading cascade-layer order declaration (`@layer reset, ...;`). */
-  emitLayerDeclaration?: boolean
-  /** Override config `minify` for this compile. */
-  minify?: boolean
-}
-
-/** One extracted call argument. A literal `null` (`{ kind: 'value', value: null }`)
- *  is distinguishable from a non-extractable argument (`{ kind: 'missing' }`). */
-export type ExtractedArg = { kind: 'value'; value: unknown } | { kind: 'missing'; value?: undefined }
-
-export interface ExtractedConditional {
-  kind: 'conditional'
-  branches: unknown[]
-}
-
-export interface ExtractedCall {
-  category: MatchCategory
-  /** Canonical Panda name; for `p.css(...)` the property name. */
-  name: string
-  /** Local binding at the call site; namespace alias for `p.css(...)`. */
-  alias: string
-  /** One entry per source argument, in order; `length` matches arity. */
-  data: ExtractedArg[]
-  span: Span
-}
-
-/** Fine-grained JSX classification: a styled factory (`<styled.div>`), a
- *  pattern component (`<Stack>`), a recipe component (`<Button>`), or a plain
- *  configured component. */
-export type JsxKind = 'factory' | 'pattern' | 'recipe' | 'component'
-
-export interface ExtractedJsx {
-  category: MatchCategory
-  /** Origin of the component — distinguishes jsx-factory/pattern/recipe. */
-  kind: JsxKind
-  name: string
-  alias: string
-  data: Record<string, unknown>
-  span: Span
-}
-
-export interface ExtractResult {
-  calls: ExtractedCall[]
-  jsx: ExtractedJsx[]
-  diagnostics: Diagnostic[]
-}
-
-export interface TokenRefSite {
-  path: string
-  span: Span
-  range: SourceRange
-  needsCssVar: boolean
-  /** `true` when the call was `token.var(...)` rather than `token(...)`. */
-  isVar: boolean
-  resolved: boolean
-  category?: string
-}
-
-export type ComponentEntryKind = 'jsx-component' | 'jsx-pattern' | 'jsx-recipe' | 'jsx-slot-recipe'
-
-export interface ComponentEntryRef {
-  kind: ComponentEntryKind
-  name: string
-  span: Span
-  range: SourceRange
-  recipe?: string
-  slot?: string
-  pattern?: string
-}
-
-export type StyleEntryKind = 'utility' | 'condition' | 'selector' | 'recipe-variant' | 'pattern-prop' | 'unknown'
-export type StyleEntrySyntax =
-  | 'css-call'
-  | 'jsx-prop'
-  | 'jsx-style-prop'
-  | 'recipe-call'
-  | 'pattern-call'
-  | 'template-style'
-export type StyleEntryOrigin = 'local' | 'cross-file' | 'generated' | 'unknown'
-export type StyleEntryFixability = 'report-only' | 'safe'
-
-export interface StyleEntryRef {
-  kind: StyleEntryKind
-  syntax: StyleEntrySyntax
-  /** Enclosing `css()`/recipe call or JSX element; `(owner, parent path)` is one style block. */
-  owner: { kind: 'call' | 'jsx'; index: number }
-  origin: StyleEntryOrigin
-  span: Span
-  range: SourceRange
-  keySpan?: Span
-  valueSpan?: Span
-  path: string[]
-  name: string
-  canonicalName?: string
-  shorthandOf?: string
-  sourceValue: unknown
-  resolvedValue: unknown
-  fixable: StyleEntryFixability
-  /** Source span of each string leaf, so fixers can target the exact literal. */
-  valueSpans?: { value: string; span: Span }[]
-}
-
-/** An extracted call/JSX with a resolved source range, for inspection consumers. */
-export interface InspectionCall extends ExtractedCall {
-  range: SourceRange
-}
-export interface InspectionJsx extends ExtractedJsx {
-  range: SourceRange
-}
-
-export interface FileInspectionResult {
-  usages: UsageSite[]
-  diagnostics: Diagnostic[]
-  calls: InspectionCall[]
-  jsx: InspectionJsx[]
-  tokenRefs: TokenRefSite[]
-  componentEntries: ComponentEntryRef[]
-  styleEntries: StyleEntryRef[]
-}
-
-/** Author import map (`ImportMapInput`). */
-export interface ImportMapInput {
-  css?: string | string[]
-  recipe?: string | string[]
-  recipes?: string | string[]
-  pattern?: string | string[]
-  patterns?: string | string[]
-  jsx?: string | string[]
-  tokens?: string | string[]
-}
-
-/** Normalized import map for snapshots and Rust. */
-export interface ImportMapOutput {
-  css: string[]
-  recipe: string[]
-  pattern: string[]
-  jsx: string[]
-  tokens: string[]
-}
-
-/** Styled-system root string or per-category object. */
-export type ImportMapOption = string | ImportMapInput
-
-/** JSON-safe, fully-resolved Panda config the compiler consumes.
- *
- *  Intentionally opaque: function-valued options (`utilities.*.transform`, …)
- *  are already lowered to `{ kind: 'js-callback', id }` refs, with the live
- *  functions supplied via {@link CompilerOptions.callbacks}. `@pandacss/config`
- *  produces values conforming to this; the compiler never imports config.
- *  `importMap` is expanded to {@link ImportMapOutput} before Rust. */
-export type SerializedConfig = Record<string, unknown>
-
-export interface SerializedRegex {
-  kind: 'regex'
-  source: string
-  flags?: string
-}
-
-export type SerializedHookPattern =
-  | string
-  | SerializedRegex
-  | {
-      include?: Array<string | SerializedRegex>
-      exclude?: Array<string | SerializedRegex>
-    }
-
-export interface SerializedHookFilter {
-  id?: SerializedHookPattern
-  code?: {
-    include?: string | SerializedRegex
-    exclude?: string | SerializedRegex
-  }
-}
-
-export interface SerializedSourceTransformHook {
-  id: string
-  name?: string
-  filter?: SerializedHookFilter
-  hash: string
-}
-
-export interface ProjectHooks {
-  'parser:before'?: SerializedSourceTransformHook[]
-}
-
-export type ProjectCallbackKind =
-  | 'utility.transform'
-  | 'utility.values'
-  | 'pattern.transform'
-  | 'pattern.defaultValues'
-  | 'parser:before'
-
-/** Common subset of the native `TokenDictionary` / wasm `TokenDictionaryInput`. */
-export interface TokenLookup {
-  values: Record<string, string>
-  vars: Record<string, string>
-}
-
-export interface RawToken {
-  path: string
-  value: string
-  var?: string
-}
-
-export interface ColorMixResult {
-  invalid: boolean
-  value: string
-  color?: string
-}
-
-export interface TransformArgs {
-  token: ((path: string) => string | undefined) & { raw: (path: string) => RawToken | undefined }
-  raw: unknown
-  utils: {
-    colorMix(value: string): ColorMixResult
-  }
-}
-
-export interface PatternHelpers {
-  map(value: unknown, fn: (value: any) => any): unknown
-  isCssUnit(value: unknown): boolean
-  isCssVar(value: unknown): boolean
-  isCssFunction(value: unknown): boolean
-}
-
-export type UtilityValuesTheme = (category: string) => Record<string, string> | undefined
-export type UtilityValuesCallback = (theme: UtilityValuesTheme) => Record<string, string> | string[] | undefined
-export type UtilityTransformCallback = (value: string, args: TransformArgs) => unknown
-export type PatternTransformCallback = (props: Record<string, any>, helpers: PatternHelpers) => unknown
-export type PatternDefaultValuesCallback = (props: Record<string, any>) => unknown
-export interface SourceTransformArgs {
-  filePath: string
-  content: string
-  original?: string
-}
-export type SourceTransformCallback = (args: SourceTransformArgs) => string | void
-
-export interface ProjectCallbackMap {
-  'utility.transform': Record<string, UtilityTransformCallback>
-  'utility.values': Record<string, UtilityValuesCallback>
-  'pattern.transform': Record<string, PatternTransformCallback>
-  'pattern.defaultValues': Record<string, PatternDefaultValuesCallback>
-  'parser:before': Record<string, SourceTransformCallback>
-}
-
-export type ProjectCallbacks = Partial<ProjectCallbackMap>
-
-/** Serialized config paired with its live callbacks, as produced by the host's
- *  `createConfigSnapshot`. */
-export interface ConfigSnapshot {
-  config: SerializedConfig
-  callbacks?: ProjectCallbacks
-  hooks?: ProjectHooks
-}
-
-export interface CompilerOptions {
-  /** Defaults to `true`; lets `token('…')` and `import { x } from './tokens'` fold. */
-  crossFile?: boolean
-  callbacks?: ProjectCallbacks
-  hooks?: ProjectHooks
-}
-
-/** In-memory filesystem handle, exposed as `Compiler.fs` on the wasm binding
- *  (the browser has no real FS); `undefined` on native. */
-export interface CompilerFileSystem {
-  /** Add or replace a source file in the in-memory filesystem. */
-  addFile(path: string, content: string): void
-  /** Remove a file from the in-memory filesystem. Returns `true` when it existed. */
-  removeFile(path: string): boolean
-  /** Read a file from the in-memory filesystem. Returns `undefined` when missing. */
-  readFile(path: string): string | undefined
-  /** `true` when `path` exists as a file or directory. */
-  exists(path: string): boolean
-  /** Number of files currently staged in the in-memory filesystem. */
-  fileCount(): number
-}
-
-// ---------------------------------------------------------------------------
-// Build info — portable design-system encoder state (`panda.buildinfo.json`)
-// ---------------------------------------------------------------------------
-
-/** A value-interned atom: `p`rop / `v`alue / `c`onditions reference the
- *  `BuildInfo.strings` table; `v` is a bare index (string),
- *  `{ t, v }` (token path + resolved value), `{ n }` (number, drives px), a
- *  boolean, or null. `c`/`i` omitted when empty/false. */
-export interface BuildAtom {
-  p: number
-  v: number | { t: number; v: number } | { n: number } | boolean | null
-  c?: number[]
-  i?: boolean
-}
-
-/** An interned recipe / slot-recipe style group, mirroring the engine's
- *  `RecipeStyleGroupSnapshot`: `r`ecipe name, optional `slot`, `cls` class name
- *  (all string indices), `cond`itions, and `entries` ({@link BuildAtom}). */
-export interface BuildRecipeGroup {
-  r: number
-  slot?: number
-  cls: number
-  cond?: number[]
-  entries: BuildAtom[]
-}
-
-/** Interned recipe state. `base`/`variants` are addressed by a combined index
- *  (base first, variants continuing) — what `modules[].recipes` references;
- *  `atomic` styles hydrate wholesale (they're recipe-wide). */
-export interface BuildRecipes {
-  base?: BuildRecipeGroup[]
-  variants?: BuildRecipeGroup[]
-  compounds?: BuildRecipeGroup[]
-  atomic?: BuildAtom[]
-}
-
-/** Per-module provenance: indices into the artifact's `atoms` / `recipes` that a
- *  given library module uses. Drives module-filtered (tree-shaken) hydration. */
-export interface BuildModuleEntry {
-  atoms?: number[]
-  recipes?: number[]
-}
-
-/** `panda.buildinfo.json` — a design-system library's serialized encoder state.
- *  A consumer hydrates it (optionally per-module, for tree-shaking) instead of
- *  re-extracting the library's components. */
-export interface BuildInfo {
-  schemaVersion: number
-  /** Peer Panda version range the artifact was built against (collision guard);
-   *  author-supplied via `create({ panda })`. */
-  panda: string
-  /** The producing engine's fingerprint of its output-affecting config (collision
-   *  guard). Stable across machines; a consumer compares it against its own
-   *  `compiler.buildInfo.configFingerprint`. */
-  configFingerprint: string
-  /** Intern table — every prop / condition / value string, referenced by index. */
-  strings: string[]
-  atoms: BuildAtom[]
-  /** Interned config recipe / slot-recipe styles; absent when the library has
-   *  none. Inline `cva`/`sva` and patterns are _not_ here — they decompose to
-   *  `atoms` in their module and travel that way. */
-  recipes?: BuildRecipes
-  /** Published module key → indices into `atoms` / `recipes`, for module-filtered
-   *  hydration. Keys are importable module ids (e.g. `"button"` for
-   *  `@acme/ds/button`). */
-  modules: Record<string, BuildModuleEntry>
-  /** Exported component name → module key, so a consumer can resolve a barrel
-   *  import (`import { Button } from '@acme/ds'`) to the module whose styles it
-   *  needs (e.g. the one whose recipe `Button` consumes). Emitted by the engine
-   *  for modules that contribute styles; absent when empty. Cross-file re-exports
-   *  (`export { X } from './y'`, `export * from './y'`) are resolved project-side
-   *  before serialization. */
-  exports?: Record<string, string>
-}
-
-/** Why an external build info is incompatible with the consuming compiler.
- *  `schemaVersion` is decided by the engine (`validate`); `pandaRange` is a
- *  host-layer verdict (it knows the running Panda version). */
-export type BuildInfoIncompatibility = 'schemaVersion' | 'pandaRange'
-
-/** Discriminated result so `ok: true` narrows away `reason`. */
-export type BuildInfoCompatibility = { ok: true } | { ok: false; reason: BuildInfoIncompatibility }
-
-export interface BuildInfoCreateOptions {
-  /** Peer Panda version range to stamp into the artifact (the library author's
-   *  published compatibility range). */
-  panda: string
-}
-
-export interface BuildInfoHydrateOptions {
-  /** Stable name for the source design system — keys the hydrated atoms so a
-   *  later hydrate of the same name replaces cleanly (e.g. the package name). */
-  name: string
-  /** Restrict hydration to these module keys (the consumer's imported
-   *  components) so only their CSS emits — tree-shaking. Omit to hydrate all. */
-  only?: string[]
-}
-
-/** Discriminated result. On success, `modules` is the set actually hydrated
- *  (the requested `only` intersected with the artifact's modules, or all of
- *  them). On failure it's empty and `reason` says why. */
-export type BuildInfoHydrateResult =
-  | { ok: true; modules: string[] }
-  | { ok: false; reason: BuildInfoIncompatibility; modules: [] }
-
-/** Produce, validate, and hydrate `panda.buildinfo.json` — a design system's
- *  portable encoder state. Accessed as `compiler.buildInfo`. */
-export interface BuildInfoApi {
-  /** This compiler's own config fingerprint — the value `create` stamps as
-   *  `configFingerprint`. A consumer compares it against an artifact's
-   *  `configFingerprint` to tell whether both sides were built with the same
-   *  output-affecting config. */
-  readonly configFingerprint: string
-
-  /** Serialize this project's encoded atoms + recipes into a portable build info,
-   *  with per-module provenance — the producer side (`panda buildinfo`). */
-  create(options: BuildInfoCreateOptions): BuildInfo
-
-  /** Check an external build info's wire compatibility with this compiler
-   *  (`schemaVersion`) without mutating anything. The Panda peer-range verdict is
-   *  layered on by the host, which knows the running version. */
-  validate(buildInfo: BuildInfo): BuildInfoCompatibility
-
-  /** Resolve which library modules a set of barrel-imported export names touch,
-   *  via the build info's `exports` map — the bridge from "what the app imports"
-   *  to the `only` filter. Subpath imports are already module keys (pass them
-   *  straight to `hydrate`); a namespace import means "all modules" (omit
-   *  `only`). Unknown names are ignored. */
-  modulesFor(buildInfo: BuildInfo, exportNames: string[]): string[]
-
-  /** Hydrate an external build info into this project — the consumer side
-   *  (`designSystem`). Validates first; an incompatible artifact is a no-op
-   *  ({@link BuildInfoHydrateResult.ok} `false`) the caller handles by
-   *  re-extracting the library's source. */
-  hydrate(buildInfo: BuildInfo, options: BuildInfoHydrateOptions): BuildInfoHydrateResult
-}
-
-// ---------------------------------------------------------------------------
-// Design-system manifest — `panda.lib.json` (the `designSystem` field)
-// ---------------------------------------------------------------------------
-
-/** Published import specifiers per category, so a library's compiled JSX keeps
- *  importing from the DS (`@acme/ds/css`), not the consumer's `styled-system`. */
-export interface DesignSystemManifestImportMap {
-  css?: string
-  recipes?: string
-  patterns?: string
-  jsx?: string
-  tokens?: string
-}
-
-/** Host-supplied manifest fields — everything but the engine-stamped
- *  `schemaVersion`. Paths are relative to the manifest's own directory. */
-export interface DesignSystemManifestInput {
-  name: string
-  /** Informational; powers the drift receipt, never enforced. */
-  version?: string
-  /** Peer Panda range the consumer's Panda must satisfy. */
-  panda: string
-  preset: string
-  buildInfo: string
-  importMap?: DesignSystemManifestImportMap
-  /** This library's own parent design system — the chain link. Absent at a root. */
-  designSystem?: string
-  /** Re-extract fallback globs when build info can't be hydrated (version skew). */
-  files?: string[]
-}
-
-/** `panda.lib.json` — the published manifest: host input plus the engine-stamped
- *  `schemaVersion`. */
-export interface DesignSystemManifest extends DesignSystemManifestInput {
-  schemaVersion: number
-}
-
-/** Why a manifest is incompatible: a `schemaVersion` wire mismatch, or the
- *  consumer's Panda major outside the manifest's `panda` range (`pandaRange`). */
-export type DesignSystemManifestIncompatibility = 'schemaVersion' | 'pandaRange'
-
-/** Discriminated result so `ok: true` narrows away `reason`. */
-export type DesignSystemManifestCompatibility =
-  | { ok: true }
-  | { ok: false; reason: DesignSystemManifestIncompatibility }
-
-/** Options for {@link DesignSystemApi.validate}. */
-export interface DesignSystemValidateOptions {
-  /** The consumer's running Panda version. When given, `validate` also checks
-   *  its major against the manifest's `panda` range — a `pandaRange` failure on
-   *  mismatch. Omit to check `schemaVersion` only. */
-  pandaVersion?: string
-}
-
-/** Raw chain plan the binding returns — a status-tagged value the namespace
- *  reshapes into {@link DesignSystemChainResult}. */
-export type DesignSystemChainPlan = { status: 'ordered'; order: string[] } | { status: 'cycle'; cycle: string[] }
-
-/** Outcome of `resolveChain`: the deduped root-first merge/hydrate `order`, or
- *  the `cycle` loop path on failure. */
-export type DesignSystemChainResult = { ok: true; order: string[] } | { ok: false; reason: 'cycle'; cycle: string[] }
-
-/** Inputs to `load` — the library's already-read build info plus the consumer's
- *  imports from the design system, for tree-shaking. */
-export interface DesignSystemLoadOptions {
-  /** The library's portable encoder state (its `panda.buildinfo.json`), already
-   *  read off disk by the host. */
-  buildInfo: BuildInfo
-  /** Export names the consumer imports from the design system; resolved to
-   *  modules via the build info's `exports` so only their CSS emits. Omit to
-   *  hydrate every module (e.g. a namespace import). */
-  imports?: string[]
-  /** The consumer's running Panda version, forwarded to `validate` so `load`
-   *  also gates on the `panda` range. See {@link DesignSystemValidateOptions}. */
-  pandaVersion?: string
-}
-
-/** Outcome of `load`: the hydrated `modules` keyed under the manifest name, or a
- *  `reason` (wire mismatch / incompatible build info) the host falls back on by
- *  re-extracting the manifest's `files`. */
-export type DesignSystemLoadResult =
-  | { ok: true; name: string; modules: string[] }
-  | { ok: false; reason: DesignSystemManifestIncompatibility | BuildInfoIncompatibility; modules: [] }
-
-/** Produce, validate, and load `panda.lib.json`. Accessed as
- *  `compiler.designSystem`. */
-export interface DesignSystemApi {
-  /** The manifest wire-format version this compiler reads/writes. */
-  readonly schemaVersion: number
-
-  /** Build a manifest from host-supplied fields, stamping the schema version —
-   *  the producer side (`panda lib`). */
-  create(input: DesignSystemManifestInput): DesignSystemManifest
-
-  /** Check a manifest's `schemaVersion` against this compiler, plus the `panda`
-   *  major against `options.pandaVersion` when supplied — so `ok: true` means
-   *  safe to use, not just a matching wire version. */
-  validate(manifest: DesignSystemManifest, options?: DesignSystemValidateOptions): DesignSystemManifestCompatibility
-
-  /** Consumer side (`designSystem: '@acme/ds'`): validate the manifest and
-   *  hydrate the modules the consumer's imports touch, under the manifest name.
-   *  The build-info half only — preset + `importMap` merge in the host. */
-  load(manifest: DesignSystemManifest, options: DesignSystemLoadOptions): DesignSystemLoadResult
-
-  /** Order already-read manifests by their parent (`designSystem`) links into a
-   *  deduped, root-first plan — the composition case. Catches cycles. */
-  resolveChain(manifests: DesignSystemManifest[]): DesignSystemChainResult
-}
-
-/** A configured, long-lived compiler: built once from a serialized config, fed
- *  source files, drained to CSS via `compile()`. Identical on native and wasm —
- *  only construction differs (native sync; wasm async + populates `fs`). */
-export interface Compiler {
-  /** Environment-specific filesystem handle. Present only on wasm/browser. */
-  readonly fs?: CompilerFileSystem
-
-  // Config and introspection
-  /** Serialized config snapshot this compiler was created with. */
-  config(): SerializedConfig
-  /** Resolved cascade-layer names (config overrides merged over defaults) — so
-   *  the host can recognize the user's `@layer …;` directive. */
-  layers(): LayerNames
-  /** Whether `css` has Panda's cascade-layer declaration (`@layer reset, base, …;`),
-   *  marking it as the stylesheet root a bundler injects the compiled CSS into. */
-  hasLayerDeclaration(css: string): boolean
-  /** Tooling introspection snapshot — read once, index on the host. */
-  spec(): Spec
-  /** Source globs + their static base dirs (for the host watcher). */
-  sources(): SourceEntry[]
-  /** Config validation diagnostics captured at construction time. */
-  diagnostics(): Diagnostic[]
-
-  // File discovery and parsing
-  /** Source paths matching the config's `include`/`exclude` (overridable) via
-   *  the filesystem engine — for discovery and host watch lists. */
-  scan(options?: ScanOptions): string[]
-  /** Real on-disk path (absolute, symlinks followed) via the filesystem engine,
-   *  so two paths to the same file compare equal. Returns the input if unresolved. */
-  realpath(path: string): string
-  /** Resolve `path` against the compiler host cwd. Absolute paths are returned unchanged. */
-  resolvePath(path: string, cwd?: string): string
-  /** Join path segments using the compiler host path semantics. */
-  joinPath(parts: string[]): string
-  /** Parent path using the compiler host path semantics. */
-  dirname(path: string): string
-  /** Whether `path` is a source file the project extracts from — its `cwd`-relative
-   *  form matches the configured `include`/`exclude` globs. For routing watch events. */
-  isSourceFile(path: string): boolean
-  /** Read + parse paths returned from `scan()`, returning one report per file
-   *  that was successfully read and parsed. */
-  parseFiles(paths: string[]): ParseFileReport[]
-  /** Read + parse a path from the compiler filesystem. */
-  parseFile(path: string): ParseFileReport
-  /** Parse provided source text for `path`. */
-  parseFileSource(path: string, source: string): ParseFileReport
-  /** Re-parse `path` only if already known; `true` when present. Filter watch
-   *  events through this to ignore unrelated files. */
-  refreshFile(path: string): boolean
-  /** Re-parse provided source text for `path` only if already known. */
-  refreshFileSource(path: string, source: string): boolean
-  /** Drop one known file's atoms, recipes, and diagnostics from project state. */
-  removeFile(path: string): boolean
-  /** Drop all parsed file state while keeping the compiled config. */
-  clear(): void
-
-  // CSS output
-  /** Compile the current project state into a complete stylesheet. */
-  compile(options?: CompileOptions): CompileOutput
-  /** CSS for the named layers only, concatenated in order — sliced in Rust so
-   *  byte offsets stay valid. Pairs with `layers()` (names). Backs
-   *  `cssgen --minimal`. */
-  getLayerCss(options: LayerCssOptions): CompileOutput
-  /** The stylesheet as a set of writable files (one per layer + per recipe,
-   *  plus `recipes.css` / `styles.css` index files) — backs `cssgen --splitting`.
-   *  The host writes each `path -> code`, the same model as `writeArtifacts`. */
-  getSplitCss(options?: SplitCssOptions): CssFile[]
-
-  /** Build-info distribution — produce, validate, and hydrate a design system's
-   *  serialized encoder state. See {@link BuildInfoApi}. */
-  readonly buildInfo: BuildInfoApi
-
-  /** Design-system manifest — produce + validate `panda.lib.json`. See
-   *  {@link DesignSystemApi}. */
-  readonly designSystem: DesignSystemApi
-
-  // Codegen artifacts
-  /** Generate + write artifacts under `outdir` via the platform fs (disk on
-   *  native, in-memory on wasm). Returns the written paths. */
-  writeArtifacts(options: WriteArtifactsOptions): string[]
-  /** Generate + write stylesheet CSS via the platform fs. */
-  writeCss(options: WriteCssOptions): WriteCssResult
-  /** Generate + write CSS for selected layers via the platform fs. */
-  writeLayerCss(options: WriteLayerCssOptions): WriteCssResult
-  /** Generate + write split stylesheet files under `outdir` via the platform fs. */
-  writeSplitCss(options: WriteSplitCssOptions): WriteFilesResult
-  /** Generate all codegen artifacts in memory without writing them. */
-  generateArtifacts(options?: GenerateArtifactOptions): CodegenArtifact[]
-  /** Generate a single codegen artifact by id. */
-  generateArtifact(id: CodegenArtifactId, options?: GenerateArtifactOptions): CodegenArtifact | undefined
-  /** Generate only artifacts affected by the provided config dependency set. */
-  generateAffectedArtifacts(dependencies: CodegenDependency[], options?: GenerateArtifactOptions): CodegenArtifact[]
-
-  // Tooling queries
-  /** Stateless source extraction — returns raw calls/jsx, registers nothing. */
-  extractFileSource(path: string, source: string): ExtractResult
-  /** Stateless source inspection for classified Panda usage sites and
-   *  file-local extraction diagnostics. */
-  inspectFileSource(path: string, source: string): FileInspectionResult
-  /** Resolve selector and CSS metadata for one utility prop/value pair. */
-  resolveUtilityValue(input: ResolveUtilityValueInput): ResolvedUtilityValue | null
-  /** Tokens that carry a hardcoded value, ranked (safe equivalents first). */
-  suggestTokens(prop: string, value: string): TokenSuggestion[]
-
-  // Project state views
-  /** Deduplicated atoms across all currently parsed files. */
-  atoms(): Atom[]
-  /** Inline and config `cva()` recipes in stable `(file, span)` order. */
-  recipes(): RecipeEntry[]
-  /** Inline and config `sva()` slot recipes in stable `(file, span)` order. */
-  slotRecipes(): RecipeEntry[]
-  /** Encoded recipe styles accumulated from parsed files. */
-  encodedRecipes(): EncodedRecipeStyles
-  /** Static pattern atoms expanded from `staticCss.patterns`. */
-  staticPatternAtoms(): StaticPatternResult
-  /** Read-only view for one parsed file, or `null` when unknown. */
-  getFile(path: string): ParsedFileView | null
-  /** Stable `(path, source hash)` manifest for known files. */
-  fileManifest(): CompileFileManifest[]
-  /** Cheap aggregate counts for the current project state. */
-  summary(): ProjectSummary
-  /** `true` when no parsed file or accumulated output state is present. */
-  isEmpty(): boolean
-}
+export * from './types/diagnostics'
+export * from './types/extraction'
+export * from './types/tokens'
+export * from './types/recipes'
+export * from './types/css'
+export * from './types/spec'
+export * from './types/codegen'
+export * from './types/parsing'
+export * from './types/config'
+export * from './types/callbacks'
+export * from './types/build-info'
+export * from './types/compiler'
