+++ packages/compiler-shared/src/types/recipes.ts
@@ -1,0 +1,27 @@
+import type { Atom } from './extraction'
+
+/** `recipe` is the serialized shape of `pandacss_recipes::Recipe`/`SlotRecipe`. */
+export interface RecipeEntry {
+  file: string
+  spanStart: number
+  recipe: unknown
+}
+
+export interface RecipeStyleEntry {
+  prop: string
+  value: string | number | boolean | null
+  conditions: string[]
+}
+
+export interface RecipeStyleGroup {
+  recipe: string
+  slot?: string | null
+  className: string
+  entries: RecipeStyleEntry[]
+}
+
+export interface EncodedRecipeStyles {
+  base: RecipeStyleGroup[]
+  variants: RecipeStyleGroup[]
+  atomic: Atom[]
+}
